Bug 1209961 - Rename fonts to bidi-proof names.
This commit removes any reference to end/beginning semantics on directional
icons: left is the beginning and right is the end only for LTR languages. RTL
languages follow the exact opposite semantic. Instead, we keep the left/right in
names for these icons, and let the user of these fonts deals with bidi csss
rules accordingly.